# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Heat Pump FDD API",
    "description": "Diagnose vapour-compression heat-pump faults from cycle features or simulated conditions.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/stats": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "One simulated point for the stats cards",
        "operationId": "getStats",
        "parameters": [
          {
            "name": "fault_type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FaultLabel",
              "default": "Condenser_Fouling"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/activity": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Short live trace for the activity strip",
        "operationId": "getActivity",
        "parameters": [
          {
            "name": "fault_type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FaultLabel",
              "default": "Condenser_Fouling"
            }
          },
          {
            "name": "inject_at",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 40,
              "minimum": 1,
              "default": 7,
              "title": "Inject At"
            }
          },
          {
            "name": "n_points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 10,
              "default": 14,
              "title": "N Points"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/explore/distribution": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Per-class box stats for one feature",
        "operationId": "getDistribution",
        "parameters": [
          {
            "name": "feature",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "superheat",
              "title": "Feature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/explore/scatter": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Stratified scatter sample of two columns",
        "operationId": "getScatter",
        "parameters": [
          {
            "name": "x",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "delta_T_evap",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "delta_T_cond",
              "title": "Y"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 60,
              "default": 720,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScatterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityResponse": {
        "properties": {
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          },
          "inject_at": {
            "type": "integer",
            "title": "Inject At"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/LiveTracePoint"
            },
            "type": "array",
            "title": "Trace"
          }
        },
        "type": "object",
        "required": [
          "fault_type",
          "inject_at",
          "trace"
        ],
        "title": "ActivityResponse"
      },
...
      "BoxStats": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "min": {
            "type": "number",
            "title": "Min"
          },
          "q1": {
            "type": "number",
            "title": "Q1"
          },
          "median": {
            "type": "number",
            "title": "Median"
          },
          "q3": {
            "type": "number",
            "title": "Q3"
          },
          "max": {
            "type": "number",
            "title": "Max"
          },
          "mean": {
            "type": "number",
            "title": "Mean"
          }
        },
        "type": "object",
        "required": [
          "name",
          "min",
          "q1",
          "median",
          "q3",
          "max",
          "mean"
        ],
        "title": "BoxStats"
      },
...
      "DistributionResponse": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/BoxStats"
            },
            "type": "array",
            "title": "Boxes"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "boxes"
        ],
        "title": "DistributionResponse"
      },
...
      "FaultLabel": {
        "type": "string",
        "enum": [
          "Normal",
          "Condenser_Fouling",
          "Evaporator_Fouling",
          "Refrigerant_Undercharge",
          "Refrigerant_Overcharge",
          "Condenser_Fan_Fault",
          "Evaporator_Fan_Fault"
        ],
        "title": "FaultLabel",
        "description": "Labels the trained classifier can emit (and inject)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LiveTracePoint": {
        "properties": {
          "t": {
            "type": "integer",
            "minimum": 0.0,
            "title": "T"
          },
          "injected_fault": {
            "type": "string",
            "title": "Injected Fault"
          },
          "severity": {
            "type": "number",
            "title": "Severity"
          },
          "predicted": {
            "type": "string",
            "title": "Predicted"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "P_evap": {
            "type": "number",
            "title": "P Evap"
          },
          "P_cond": {
            "type": "number",
            "title": "P Cond"
          },
          "T_discharge": {
            "type": "number",
            "title": "T Discharge"
          },
          "superheat": {
            "type": "number",
            "title": "Superheat"
          },
          "subcooling": {
            "type": "number",
            "title": "Subcooling"
          },
          "COP": {
            "type": "number",
            "title": "Cop"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "t",
          "injected_fault",
          "severity",
          "predicted",
          "confidence",
          "P_evap",
          "P_cond",
          "T_discharge",
          "superheat",
          "subcooling",
          "COP"
        ],
        "title": "LiveTracePoint"
      },
...
      "ScatterPoint": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "fault_type"
        ],
        "title": "ScatterPoint"
      },
      "ScatterResponse": {
        "properties": {
          "x": {
            "type": "string",
            "title": "X"
          },
          "y": {
            "type": "string",
            "title": "Y"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ScatterPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "points"
        ],
        "title": "ScatterResponse"
      },
...
      "StatsResponse": {
        "properties": {
          "fault_type": {
            "type": "string",
            "title": "Fault Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Probabilities"
          },
          "COP": {
            "type": "number",
            "title": "Cop"
          },
          "P_cond": {
            "type": "number",
            "title": "P Cond"
          },
          "P_evap": {
            "type": "number",
            "title": "P Evap"
          },
          "T_discharge": {
            "type": "number",
            "title": "T Discharge"
          },
          "superheat": {
            "type": "number",
            "title": "Superheat"
          },
          "subcooling": {
            "type": "number",
            "title": "Subcooling"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "fault_type",
          "label",
          "confidence",
          "probabilities",
          "COP",
          "P_cond",
          "P_evap",
          "T_discharge",
          "superheat",
          "subcooling",
          "model"
        ],
        "title": "StatsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}